replace all viewers with a single viewer
