--- src/HashBus.Viewer.MostMentioned/packages.config
@@ -1,8 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="ColoredConsole" version="0.4.0" targetFramework="net451" />
-  <package id="GitVersionTask" version="3.6.2" targetFramework="net451" developmentDependency="true" />
-  <package id="LiteGuard" version="0.10.0" targetFramework="net451" />
-  <package id="Newtonsoft.Json" version="7.0.1" targetFramework="net451" />
-  <package id="RestSharp" version="105.2.3" targetFramework="net451" />
-</packages>
