--- src/HashBus.Viewer.Multiple/HashBus.Viewer.Multiple.csproj
@@ -1,128 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{8E09CB51-03C5-413A-8570-1A121A01B0A4}</ProjectGuid>
-    <OutputType>Exe</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>HashBus.Viewer.Multiple</RootNamespace>
-    <AssemblyName>HashBus.Viewer.Multiple</AssemblyName>
-    <TargetFrameworkVersion>v4.5.1</TargetFrameworkVersion>
-    <FileAlignment>512</FileAlignment>
-    <SolutionDir Condition="$(SolutionDir) == '' Or $(SolutionDir) == '*Undefined*'">..\</SolutionDir>
-    <TargetFrameworkProfile />
-    <NuGetPackageImportStamp>
-    </NuGetPackageImportStamp>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <UseVSHostingProcess>false</UseVSHostingProcess>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <UseVSHostingProcess>false</UseVSHostingProcess>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-  </PropertyGroup>
-  <PropertyGroup>
-    <StartupObject />
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="ColoredConsole, Version=0.0.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\..\packages\ColoredConsole.0.4.0\lib\net45\ColoredConsole.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
-    <Reference Include="LiteGuard, Version=0.10.0.0, Culture=neutral, PublicKeyToken=d38e0b7ae24b08a1, processorArchitecture=MSIL">
-      <HintPath>..\..\packages\LiteGuard.0.10.0\lib\net35\LiteGuard.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
-    <Reference Include="Newtonsoft.Json, Version=7.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
-      <HintPath>..\..\packages\Newtonsoft.Json.7.0.1\lib\net45\Newtonsoft.Json.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
-    <Reference Include="RestSharp, Version=105.2.3.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\..\packages\RestSharp.105.2.3\lib\net451\RestSharp.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
-    <Reference Include="System" />
-    <Reference Include="System.ComponentModel.Composition" />
-    <Reference Include="System.configuration" />
-    <Reference Include="System.Core" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="..\HashBus.Viewer.MostHashtagged\MostHashtaggedLeaderBoardViewFactory.cs">
-      <Link>MostHashtaggedLeaderBoardViewFactory.cs</Link>
-    </Compile>
-    <Compile Include="..\HashBus.Viewer.MostMentioned\ColorTokenExtensions.cs">
-      <Link>ColorTokenExtensions.cs</Link>
-    </Compile>
-    <Compile Include="..\HashBus.Viewer.MostMentioned\IRunAsync.cs">
-      <Link>IRunAsync.cs</Link>
-    </Compile>
-    <Compile Include="..\HashBus.Viewer.MostMentioned\IService.cs">
-      <Link>IService.cs</Link>
-    </Compile>
-    <Compile Include="..\HashBus.Viewer.MostMentioned\LeaderboardService.cs">
-      <Link>LeaderboardService.cs</Link>
-    </Compile>
-    <Compile Include="..\HashBus.Viewer.MostMentioned\LeaderboardView.cs">
-      <Link>LeaderboardView.cs</Link>
-    </Compile>
-    <Compile Include="..\HashBus.Viewer.MostMentioned\MostMentionedLeaderBoardViewFactory.cs">
-      <Link>MostMentionedLeaderBoardViewFactory.cs</Link>
-    </Compile>
-    <Compile Include="..\HashBus.Viewer.MostRetweeted\MostRetweetedLeaderBoardViewFactory.cs">
-      <Link>MostRetweetedLeaderBoardViewFactory.cs</Link>
-    </Compile>
-    <Compile Include="..\HashBus.Viewer.TopRetweeters\TopRetweetersLeaderBoardViewFactory.cs">
-      <Link>TopRetweetersLeaderBoardViewFactory.cs</Link>
-    </Compile>
-    <Compile Include="..\HashBus.Viewer.TopTweetersRetweeters\TopTweetersRetweetersLeaderBoardViewFactory.cs">
-      <Link>TopTweetersRetweetersLeaderBoardViewFactory.cs</Link>
-    </Compile>
-    <Compile Include="..\HashBus.Viewer.TopTweeters\TopTweetersLeaderBoardViewFactory.cs">
-      <Link>TopTweetersLeaderBoardViewFactory.cs</Link>
-    </Compile>
-    <Compile Include="..\HashBus.WebApi\HashtagEntry.cs">
-      <Link>HashtagEntry.cs</Link>
-    </Compile>
-    <Compile Include="..\HashBus.WebApi\IEntry.cs">
-      <Link>IEntry.cs</Link>
-    </Compile>
-    <Compile Include="..\HashBus.WebApi\Leaderboard.cs">
-      <Link>Leaderboard.cs</Link>
-    </Compile>
-    <Compile Include="..\HashBus.WebApi\UserEntry.cs">
-      <Link>UserEntry.cs</Link>
-    </Compile>
-    <Compile Include="App.cs" />
-    <Compile Include="Program.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="app.config">
-      <SubType>Designer</SubType>
-    </None>
-    <None Include="packages.config" />
-  </ItemGroup>
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-  <Import Project="..\..\packages\GitVersionTask.3.6.2\build\dotnet\GitVersionTask.targets" Condition="Exists('..\..\packages\GitVersionTask.3.6.2\build\dotnet\GitVersionTask.targets')" />
-  <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
-    <PropertyGroup>
-      <ErrorText>This project references NuGet package(s) that are missing on this computer. Use NuGet Package Restore to download them.  For more information, see http://go.microsoft.com/fwlink/?LinkID=322105. The missing file is {0}.</ErrorText>
-    </PropertyGroup>
-    <Error Condition="!Exists('..\..\packages\GitVersionTask.3.6.2\build\dotnet\GitVersionTask.targets')" Text="$([System.String]::Format('$(ErrorText)', '..\..\packages\GitVersionTask.3.6.2\build\dotnet\GitVersionTask.targets'))" />
-  </Target>
-</Project>
