--- src/HashBus.Viewer.TopRetweeters/App.cs
@@ -1,25 +1,0 @@
-﻿namespace HashBus.Viewer.TopRetweeters
-{
-    using System.Threading.Tasks;
-    using RestSharp;
-
-    class App
-    {
-        public static async Task RunAsync(
-            string webApiBaseUrl,
-            string track,
-            int refreshInterval,
-            bool showPercentages,
-            int verticalPadding,
-            int horizontalPadding)
-        {
-            await TopRetweetersLeaderBoardViewFactory.Create(
-                    track, refreshInterval,
-                    showPercentages,
-                    verticalPadding,
-                    horizontalPadding,
-                    new RestClient(webApiBaseUrl))
-                .RunAsync();
-        }
-    }
-}
