--- src/HashBus.Viewer.TopTweetersRetweeters/app.config
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8" standalone="yes"?>
-<configuration>
-  <appSettings>
-    <add key="WebApiBaseUrl" value="http://localhost:13373" />
-    <add key="Track" value="#Simulated" />
-    <add key="refreshInterval" value="10000" />
-    <add key="ShowPercentages" value="true" />
-    <add key="VerticalPadding" value="1" />
-    <add key="HorizontalPadding" value="8" />
-  </appSettings>
-</configuration>
