+++ src/HashBus.Viewer/IService.cs
