+++ src/HashBus.Viewer/MostRetweetedLeaderBoardViewFactory.cs
@@ -1,10 +1,10 @@
-﻿namespace HashBus.Viewer.TopTweetersRetweeters
+﻿namespace HashBus.Viewer
 {
     using ColoredConsole;
     using HashBus.WebApi;
     using RestSharp;
 
-    static class TopTweetersRetweetersLeaderBoardViewFactory
+    static class MostRetweetedLeaderBoardViewFactory
     {
         public static LeaderboardView<UserEntry> Create(
             string track,
@@ -17,14 +17,14 @@
             return new LeaderboardView<UserEntry>(
                 track,
                 refreshInterval,
-                new LeaderboardService<UserEntry>(client, "/top-tweeters-retweeters/{0}"),
+                new LeaderboardService<UserEntry>(client, "/most-retweeted/{0}"),
                 showPercentages,
                 verticalPadding,
                 horizontalPadding,
                 (entry1, entry2) => entry1.Id == entry2.Id,
                 entry => new[] { $" {entry.Name}".White(), $" @{entry.ScreenName}".Cyan(), },
-                "Top Tweeters/Retweeters",
-                "tweets/retweets");
+                "Most Retweeted",
+                "retweets");
         }
     }
 }
