+++ src/HashBus.Viewer/Program.cs
@@ -1,7 +1,8 @@
-﻿namespace HashBus.Viewer.TopRetweeters
+﻿namespace HashBus.Viewer
 {
     using System;
     using System.Configuration;
+    using System.Linq;
 
     class Program
     {
@@ -13,10 +14,23 @@
             var showPercentages = bool.Parse(ConfigurationManager.AppSettings["ShowPercentages"]);
             var verticalPadding = int.Parse(ConfigurationManager.AppSettings["VerticalPadding"]);
             var horizontalPadding = int.Parse(ConfigurationManager.AppSettings["HorizontalPadding"]);
+            var rotateInterval = int.Parse(ConfigurationManager.AppSettings["RotateInterval"]);
+            var views = ConfigurationManager.AppSettings["Views"]
+                .Split(new[] { ',' }, StringSplitOptions.RemoveEmptyEntries)
+                .Select(view => view.Trim())
+                .ToArray();
 
             Console.Title = typeof(Program).Assembly.GetName().Name;
 
-            App.RunAsync(webApiBaseUrl, track, refreshInterval, showPercentages, verticalPadding, horizontalPadding)
+            App.RunAsync(
+                    webApiBaseUrl,
+                    track,
+                    refreshInterval,
+                    showPercentages,
+                    verticalPadding,
+                    horizontalPadding,
+                    views,
+                    rotateInterval)
                 .GetAwaiter().GetResult();
         }
     }
